added better version, changed makefile, cleaning up argparse, sampling and func

+++ Sampling.h
@@ -2,7 +2,7 @@
 #define SAMPLING_H
 #define LENS 4096
 
-void* Create_se_threads(faidx_t *seq_ref,int thread_no, int seed, size_t reads,const char* OutputName,const char* Adapt_flag,const char* Adapter_1,
+void* Create_se_threads(faidx_t *seq_ref,int thread_no, int seed, long long int reads,const char* OutputName,const char* Adapt_flag,const char* Adapter_1,
                         const char* Adapter_2,const char* OutputFormat,const char* SeqType,float BriggsParam[4],const char* Briggs_Flag,
                         const char* Sizefile,int FixedSize,int SizeDistType,int val1,int val2,
                         int qualstringoffset,const char* QualProfile1,const char* QualProfile2,int threadwriteno,
